ref: Update t_philo struct to include *config member
The recent code changes involve updating the t_philo struct to include a *config member. This change allows for easier access to the configuration settings for each philosopher. The *config member will be used in various functions throughout the codebase to access the time_to_die value for monitoring and checking if a philosopher is dead.